Feat: 사원 정보 수정 api

+++ src/main/java/com/_pi/benepick/domain/members/dto/MembersRequest.java
@@ -15,7 +15,7 @@ public class MembersRequest {
         private String deptName;
         private String name;
         private Long point;
-        private Long penaltyCnt;
+        private int penaltyCnt;
     }
 
     @Builder
